requirements.txt
discord~=2.3.2
discord.py~=2.4.0
psutil~=5.9.8
python-dotenv~=1.0.1